developmet merge fixes

+++ src/pages/Blog.jsx
@@ -3,7 +3,7 @@ import { deleteObject, ref } from 'firebase/storage'
 import React from 'react'
 import { useState } from 'react'
 import { useEffect } from 'react'
-import { useNavigate, useParams } from 'react-router-dom'
+import { Link, useNavigate, useParams } from 'react-router-dom'
 import { auth, db, storage } from '../firebase-config'
 
 function Blog({postId}) {
@@ -11,6 +11,7 @@ function Blog({postId}) {
 
     const [blogPost, setBlogPost] = useState({})
     const [author, setAuthor] = useState('')
+    const [authUser, setAuthUser] = useState(null)
 
     let navigate = useNavigate()
 
@@ -18,12 +19,18 @@ function Blog({postId}) {
         try {
             const getPath = ref(storage, blogPost.imageURL)
             const objRef = ref(storage, getPath.fullPath)
-            await deleteObject(objRef)
-            await deleteDoc(doc(db, 'posts', pid));
-            navigate('/')
-        } catch (error) {
-            if(error.message === 'storage/object-not-found') return
+            if  (objRef.fullPath === '') {
                 await deleteDoc(doc(db, 'posts', pid));
+                navigate('/')
+            } else {
+                await deleteObject(objRef)
+                await deleteDoc(doc(db, 'posts', pid));
+                navigate('/')
+            }
+        } catch (error) {
+            if(error.message === 'storage/object-not-found') {
+                await deleteDoc(doc(db, 'posts', pid));
+            }
         }
     }
 
@@ -36,10 +43,11 @@ function Blog({postId}) {
             const response = await getDoc(postRef)
             setBlogPost(response.data())
             setAuthor(response.data().author)
-            console.log(auth.currentUser.uid);
+            if(auth.currentUser === null) return
+            setAuthUser(auth.currentUser.uid)
         }
         getById()
-    },[id.postId])
+    },[id.postId, authUser])
     
   return (
     <div className='blog-post__container'>
@@ -56,9 +64,11 @@ function Blog({postId}) {
             <div className="blog-post__content__body">
                 <p>{blogPost.content}</p>
             </div>
-            {author.id === auth.currentUser.uid ? 
+            {author.id === authUser ? 
                 <div className="blog-post__actions">
-                    <button>Edit</button>
+                    <button>
+                        <Link to={`/blog/edit/${id.postId}`}>Edit</Link>
+                    </button>
                     <button onClick={() => deletePost(id.postId)}>Delete</button>
                 </div> : ''
             }
